python_ml/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
numpy==2.1.2
pandas==2.2.3
Pillow==11.0.0
geopy==2.4.1
python-multipart==0.0.20
requests==2.32.3
typing-extensions==4.12.2
